chore(NCP/Contribute): Refactor contribute cards
